Monkey patch older versions of fog
Unfortunately we're stuck using an old version of vcloud-launcher and have
been caught between version incompatibilities.  To work around this the best
approach we have is to monkey patch the fog change so we can continue to work.

The addition of the 'queued' status was added in fog-1.33.0, however we're
currently stuck on fog-1.29.